pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support_env"
version = "0.1.0"
description = "OpenEnv Support Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai"
]

[project.scripts]
server = "server.app:main"
